pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-sre-incident-env"
version = "0.1.0"
description = "SRE Incident Response Environment for OpenEnv - simulates real-world DevOps on-call scenarios"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "sre_incident_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["sre_incident_env", "sre_incident_env.server"]
package-dir = { "sre_incident_env" = ".", "sre_incident_env.server" = "server" }
